Remove event carousel and modal
