+++ src/assets/anuncios-carousel/anuncio1/README.md
@@ -1,0 +1,1 @@
+Coloca aquí las imágenes del carrusel para el anuncio 1.
