+++ src/components/HeroCarousel.tsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import ImageCarousel from './ImageCarousel';
 
 // Import all images from the hero-carousel assets folder
 const images = Object.values(
@@ -9,26 +10,10 @@ const images = Object.values(
 ) as string[];
 
 const HeroCarousel = () => {
-  const [index, setIndex] = useState(0);
-
-  useEffect(() => {
-    if (images.length <= 1) return;
-    const interval = setInterval(() => {
-      setIndex((prev) => (prev + 1) % images.length);
-    }, 4000);
-    return () => clearInterval(interval);
-  }, []);
-
   if (images.length === 0) return null;
 
   return (
-    <div className="w-full h-64 md:h-96 overflow-hidden rounded-2xl shadow-lg">
-      <img
-        src={images[index]}
-        alt={`Imagen carrusel ${index + 1}`}
-        className="w-full h-full object-cover transition-opacity duration-700"
-      />
-    </div>
+    <ImageCarousel images={images} className="w-full h-64 md:h-96 shadow-lg" />
   );
 };
